+++ tests/DocumentTest.php
@@ -85,8 +85,8 @@ class DocumentTest extends TestCase
         $this->assertInstanceOf(Document::class, $document2);
         $this->assertSame($document->type(), $document2->type());
         $this->assertNotSame($document->children(), $document2->children());
-        $this->assertCount(3, $document->children());
-        $this->assertCount(4, $document2->children());
+        $this->assertSame(3, $document->children()->size());
+        $this->assertSame(4, $document2->children()->size());
         $this->assertSame(
             $node,
             $document2->children()->get(0)->match(
@@ -127,8 +127,8 @@ class DocumentTest extends TestCase
         $this->assertInstanceOf(Document::class, $document2);
         $this->assertSame($document->type(), $document2->type());
         $this->assertNotSame($document->children(), $document2->children());
-        $this->assertCount(3, $document->children());
-        $this->assertCount(4, $document2->children());
+        $this->assertSame(3, $document->children()->size());
+        $this->assertSame(4, $document2->children()->size());
         $this->assertEquals(
             $document->children()->get(0),
             $document2->children()->get(0),