+++ tests/Translator/NodeTranslator/LinkTranslatorTest.php
@@ -14,11 +14,13 @@ class LinkTranslatorTest extends TestCase
 {
     public function testTranslate()
     {
-        $dom = new \DOMDocument;
-        $dom->loadHTML('<link href="/" rel="next" hreflang="fr"/>');
+        $dom = \Dom\HTMLDocument::createFromString(
+            '<link href="/" rel="next" hreflang="fr"/>',
+            \LIBXML_HTML_NOIMPLIED | \LIBXML_NOERROR,
+        );
 
         $link = Translator::new()(
-            $dom->childNodes->item(1)->childNodes->item(0)->childNodes->item(0),
+            $dom->childNodes->item(0),
         )->match(
             static fn($link) => $link,
             static fn() => null,
@@ -28,7 +30,7 @@ class LinkTranslatorTest extends TestCase
         $this->assertSame('/', $link->href()->toString());
         $this->assertSame('next', $link->relationship());
         $link = $link->normalize();
-        $this->assertCount(3, $link->attributes());
+        $this->assertSame(3, $link->attributes()->size());
         $this->assertSame('fr', $link->attribute('hreflang')->match(
             static fn($attribute) => $attribute->value(),
             static fn() => null,
@@ -37,11 +39,13 @@ class LinkTranslatorTest extends TestCase
 
     public function testTranslateWithoutRelationship()
     {
-        $dom = new \DOMDocument;
-        $dom->loadHTML('<link href="/" hreflang="fr"/>');
+        $dom = \Dom\HTMLDocument::createFromString(
+            '<link href="/" hreflang="fr"/>',
+            \LIBXML_HTML_NOIMPLIED | \LIBXML_NOERROR,
+        );
 
         $link = Translator::new()(
-            $dom->childNodes->item(1)->childNodes->item(0)->childNodes->item(0),
+            $dom->childNodes->item(0),
         )->match(
             static fn($link) => $link,
             static fn() => null,
@@ -51,7 +55,7 @@ class LinkTranslatorTest extends TestCase
         $this->assertSame('/', $link->href()->toString());
         $this->assertSame('related', $link->relationship());
         $link = $link->normalize();
-        $this->assertCount(3, $link->attributes());
+        $this->assertSame(3, $link->attributes()->size());
         $this->assertSame('fr', $link->attribute('hreflang')->match(
             static fn($attribute) => $attribute->value(),
             static fn() => null,
@@ -64,11 +68,13 @@ class LinkTranslatorTest extends TestCase
 
     public function testReturnNothingWhenMissingHrefAttribute()
     {
-        $dom = new \DOMDocument;
-        $dom->loadHTML('<link/>');
+        $dom = \Dom\HTMLDocument::createFromString(
+            '<link/>',
+            \LIBXML_HTML_NOIMPLIED | \LIBXML_NOERROR,
+        );
 
         $result = Translator::new()(
-            $dom->childNodes->item(1)->childNodes->item(0)->childNodes->item(0),
+            $dom->childNodes->item(0),
         )
             ->maybe()
             ->keep(Instance::of(Link::class));