+++ tests/Translator/NodeTranslator/ScriptTranslatorTest.php
@@ -13,11 +13,13 @@ class ScriptTranslatorTest extends TestCase
 {
     public function testTranslate()
     {
-        $dom = new \DOMDocument;
-        $dom->loadHTML('<script type="text/javascript">var foo = 42;</script>');
+        $dom = \Dom\HTMLDocument::createFromString(
+            '<script type="text/javascript">var foo = 42;</script>',
+            \LIBXML_HTML_NOIMPLIED | \LIBXML_NOERROR,
+        );
 
         $script = Translator::new()(
-            $dom->childNodes->item(1)->childNodes->item(0)->childNodes->item(0),
+            $dom->childNodes->item(0),
         )->match(
             static fn($script) => $script,
             static fn() => null,
@@ -29,7 +31,7 @@ class ScriptTranslatorTest extends TestCase
             '<script type="text/javascript">var foo = 42;</script>'."\n",
             $script->asContent()->toString(),
         );
-        $this->assertCount(1, $script->attributes());
+        $this->assertSame(1, $script->attributes()->size());
         $this->assertSame(
             'text/javascript',
             $script->attribute('type')->match(
@@ -37,16 +39,18 @@ class ScriptTranslatorTest extends TestCase
                 static fn() => null,
             ),
         );
-        $this->assertCount(1, $script->children());
+        $this->assertSame(1, $script->children()->size());
     }
 
     public function testTranslateWithoutCode()
     {
-        $dom = new \DOMDocument;
-        $dom->loadHTML('<script></script>');
+        $dom = \Dom\HTMLDocument::createFromString(
+            '<script></script>',
+            \LIBXML_HTML_NOIMPLIED | \LIBXML_NOERROR,
+        );
 
         $script = Translator::new()(
-            $dom->childNodes->item(1)->childNodes->item(0)->childNodes->item(0),
+            $dom->childNodes->item(0),
         )->match(
             static fn($script) => $script,
             static fn() => null,
@@ -58,7 +62,7 @@ class ScriptTranslatorTest extends TestCase
             '<script></script>'."\n",
             $script->asContent()->toString(),
         );
-        $this->assertCount(0, $script->attributes());
-        $this->assertCount(1, $script->children());
+        $this->assertSame(0, $script->attributes()->size());
+        $this->assertSame(1, $script->children()->size());
     }
 }