+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Requires PHP `8.4`
+- Requires `innmind/xml:~9.0`
 
 ## 7.0.0 - 2025-06-07