+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 8.0.0 - 2026-01-25
+
+### Changed
+
+- Requires PHP `8.4`
+- Requires `innmind/xml:~9.0`
+
 ## 7.0.0 - 2025-06-07
 
 ### Added